compression_codecs::zstd::params

Struct CParameter

pub struct CParameter(CParameter);

A compression parameter for zstd. This is a stable wrapper around zstdโs own CParameter type, to abstract over different versions of the zstd library.

See the zstd documentation for more information on these parameters.

Tuple Fields
0: CParameter
Implementations

impl CParameter

pub fn quality(level: Level) -> i32

pub fn window_log(value: u32) -> Self

Window size in bytes (as a power of two)

pub fn hash_log(value: u32) -> Self

Size of the initial probe table in 4-byte entries (as a power of two)

pub fn chain_log(value: u32) -> Self

Size of the multi-probe table in 4-byte entries (as a power of two)

pub fn search_log(value: u32) -> Self

Number of search attempts (as a power of two)

pub fn min_match(value: u32) -> Self

Minimum size of matches searched for

pub fn target_length(value: u32) -> Self

Strategy-dependent length modifier

pub fn enable_long_distance_matching(value: bool) -> Self

Enable long-distance matching mode to look for and emit long-distance references.

This increases the default window size.

pub fn ldm_hash_log(value: u32) -> Self

Size of the long-distance matching table (as a power of two)

pub fn ldm_min_match(value: u32) -> Self

Minimum size of long-distance matches searched for

pub fn ldm_bucket_size_log(value: u32) -> Self

Size of each bucket in the LDM hash table for collision resolution (as a power of two)

pub fn ldm_hash_rate_log(value: u32) -> Self

Frequency of using the LDM hash table (as a power of two)

pub fn content_size_flag(value: bool) -> Self

Emit the size of the content (default: true).

pub fn checksum_flag(value: bool) -> Self

Emit a checksum (default: false).

pub fn dict_id_flag(value: bool) -> Self

Emit a dictionary ID when using a custom dictionary (default: true).

pub fn nb_workers(value: u32) -> Self

Number of threads to spawn.

If set to 0, compression functions will block; if set to 1 or more, compression will run in background threads and flush pushes bytes through the compressor.

Panics

This parameter requires feature zstdmt to be enabled, otherwise it will cause a panic when used in ZstdEncoder::with_quality_and_params() calls.

pub fn job_size(value: u32) -> Self

Number of bytes given to each worker.

If set to 0, zstd selects a job size based on compression parameters.

pub(crate) fn as_zstd(&self) -> CParameter